requirements.txt
google-generativeai>=0.8.0
google-genai>=1.0.0
pytz
